steps.model added to rating
