+++ frontend/src/app/rating/rating.page.ts
@@ -13,7 +13,6 @@ import { ToastController } from '@ionic/angular';
 import { Subscription } from 'rxjs';
 
 import { FormControl, FormGroup } from '../common/forms/forms';
-import { hrRatingStep, hrRatingStepArr } from '../forms/model/steps.model';
 import { FormsData } from '../model/forms-data.model';
 import { StorageHandlerService } from '../services/storage-handler.service';
 
@@ -22,6 +21,7 @@ import {
   ImpressionInfo,
   RatingForm,
 } from './model/rating-form.model';
+import { hrRatingStep, hrRatingStepArr } from './model/steps.model';
 
 @Component({
   selector: 'app-rating',
